opentype/gpos_kerning_info => gpos_kerning_info (Universal profile)
This is a check for "well-designed fonts" and passes if GPOS pair kerning is present.
The OpenType spec does not require that GPOS kerning to be present (nor does it require that fonts be well designed, sadly...) so this should go on the Universal profile.

(https://github.com/fonttools/fontbakery/issues/4886)