stat_has_axis_value_tables / inconsistencies_between_fvar_STAT
The spec says: "In a variable font, it is strongly recommended that axis value tables
be included for every element of typographic subfamily names for all of the named
instances defined in the 'fvar' table."

So it's a strong recommendation but not a requirement; so not strictly an OpenType
profile check. Simon suggests we deprecate opentype/stat_has_axis_value_tables,
because the other one (inconsistencies...) potentially has a wider scope in the
future, checking back from the fvar table to the STAT table.

(https://github.com/fonttools/fontbakery/issues/4916)